requirements.txt
base58
ecdsa
pyyaml
